+++ src/lib.rs
@@ -5,6 +5,7 @@
 
 extern crate alloc;
 
+mod cow_str;
 #[cfg(feature = "std")]
 mod hashset_string;
 mod string;
@@ -12,7 +13,8 @@ mod string_non_empty;
 mod vec_string;
 
 pub use crate::{
-    string::{cow_str, option_string, str, string},
+    cow_str::cow_str,
+    string::{option_string, str, string},
     string_non_empty::{option_string_non_empty, string_non_empty},
     vec_string::vec_string,
 };